pyproject.toml
[project]
name = "sip-server-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "accelerate>=1.12.0",
  "bitsandbytes>=0.48.2",
  "faster-whisper>=1.2.1",
  "gradio>=6.1.0",
  "huggingface-hub[cli]>=0.36.0",
  "jieba>=0.42.1",
  "langid>=1.1.6",
  "openai>=2.8.1",
  "piper-tts>=1.3.0",
  "pydantic>=2.12.4",
  "pydub>=0.25.1",
  "python-dotenv>=1.2.1",
  "silero-vad>=6.2.0",
  "torchcodec>=0.9.1",
  "transformers>=4.51.3",
  "voxcpm>=1.0.5",
  "websockets>=15.0.1",
]

[tool.ruff.lint]
extend-select = ["I"]
